Move BingMapsClient, BingMapsPropertiesConfig and BingMapsWebClientConfig to separate package
